+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support for subpaths when attaching datasets. For example: `--dataset 'dataset-name:/subpath:/mount/location'`.
+- Added support for subpaths when attaching datasets. For example: `--dataset 'dataset-name:sub/path:/mount/location'`.
 
 ## [v0.15.1](https://github.com/allenai/beaker-gantry/releases/tag/v0.15.1) - 2023-05-08